TextGrabber: better deal with recursive moves
When pages are moved back and forth, the grabbers can end up in a recursive loop.

The logic that previously existed only checked the previous page ID to see whether we are in a recursive loop.

This change instead mutates an array that persists through each recursive call of the function.

The values of this array is checked to determine whether we're in a loop, and (hopefully) break out of it.